+++ Source/EventFlow/Aggregates/AggregateRoot.cs
@@ -24,6 +24,7 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
+using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.EventStores;
@@ -38,14 +39,14 @@ namespace EventFlow.Aggregates
     {
         private readonly List<IUncommittedEvent> _uncommittedEvents = new List<IUncommittedEvent>();
 
-        public TIdentity Id { get; private set; }
+        public TIdentity Id { get; }
         public int Version { get; private set; }
-        public bool IsNew { get { return Version <= 0; } }
+        public bool IsNew => Version <= 0;
         public IEnumerable<IAggregateEvent> UncommittedEvents { get { return _uncommittedEvents.Select(e => e.AggregateEvent); } }
 
         protected AggregateRoot(TIdentity id)
         {
-            if (id == null) throw new ArgumentNullException("id");
+            if (id == null) throw new ArgumentNullException(nameof(id));
             if ((this as TAggregate) == null)
             {
                 throw WrongImplementationException.With(
@@ -58,12 +59,12 @@ namespace EventFlow.Aggregates
             Id = id;
         }
 
-        protected void Emit<TEvent>(TEvent aggregateEvent, IMetadata metadata = null)
+        protected virtual void Emit<TEvent>(TEvent aggregateEvent, IMetadata metadata = null)
             where TEvent : IAggregateEvent<TAggregate, TIdentity>
         {
             if (aggregateEvent == null)
             {
-                throw new ArgumentNullException("aggregateEvent");
+                throw new ArgumentNullException(nameof(aggregateEvent));
             }
 
             var now = DateTimeOffset.Now;
@@ -73,6 +74,7 @@ namespace EventFlow.Aggregates
                     {MetadataKeys.TimestampEpoch, now.ToUnixTime().ToString()},
                     {MetadataKeys.AggregateSequenceNumber, (Version + 1).ToString()},
                     {MetadataKeys.AggregateName, GetType().Name.Replace("Aggregate", string.Empty)},
+                    {MetadataKeys.AggregateId, Id.Value}
                 };
 
             metadata = metadata == null
@@ -85,7 +87,9 @@ namespace EventFlow.Aggregates
             _uncommittedEvents.Add(uncommittedEvent);
         }
 
-        public async Task<IReadOnlyCollection<IDomainEvent>> CommitAsync(IEventStore eventStore, CancellationToken cancellationToken)
+        public async Task<IReadOnlyCollection<IDomainEvent>> CommitAsync(
+            IEventStore eventStore,
+            CancellationToken cancellationToken)
         {
             var domainEvents = await eventStore.StoreAsync<TAggregate, TIdentity>(
                 Id,
@@ -111,10 +115,7 @@ namespace EventFlow.Aggregates
         {
             if (Version > 0)
             {
-                throw new InvalidOperationException(string.Format(
-                    "Aggregate '{0}' with ID '{1}' already has events",
-                    GetType().Name,
-                    Id));
+                throw new InvalidOperationException($"Aggregate '{GetType().Name}' with ID '{Id}' already has events");
             }
 
             foreach (var aggregateEvent in aggregateEvents)
@@ -122,28 +123,27 @@ namespace EventFlow.Aggregates
                 var e = aggregateEvent as IAggregateEvent<TAggregate, TIdentity>;
                 if (e == null)
                 {
-                    throw new ArgumentException(string.Format(
-                        "Aggregate event of type '{0}' does not belong with aggregate '{1}',",
-                        aggregateEvent.GetType(),
-                        this));
+                    throw new ArgumentException($"Aggregate event of type '{aggregateEvent.GetType()}' does not belong with aggregate '{this}',");
                 }
 
                 ApplyEvent(e);
             }
         }
 
-        private void ApplyEvent(IAggregateEvent<TAggregate, TIdentity> aggregateEvent)
+        protected virtual void ApplyEvent(IAggregateEvent<TAggregate, TIdentity> aggregateEvent)
         {
             var eventType = aggregateEvent.GetType();
             if (_eventHandlers.ContainsKey(eventType))
             {
                 _eventHandlers[eventType](aggregateEvent);
             }
+            else if (_eventAppliers.Any(ea => ea.Apply((TAggregate)this, aggregateEvent))) { }
             else
             {
                 var applyMethod = GetApplyMethod(eventType);
                 applyMethod(this as TAggregate, aggregateEvent);
             }
+
             Version++;
         }
 
@@ -154,15 +154,20 @@ namespace EventFlow.Aggregates
             var eventType = typeof (TAggregateEvent);
             if (_eventHandlers.ContainsKey(eventType))
             {
-                throw new ArgumentException(string.Format(
-                    "There's already a event handler registered for the aggregate event '{0}'",
-                    eventType.Name));
+                throw new ArgumentException($"There's already a event handler registered for the aggregate event '{eventType.Name}'");
             }
             _eventHandlers[eventType] = e => handler((TAggregateEvent)e);
         }
 
+        private readonly List<IEventApplier<TAggregate, TIdentity>> _eventAppliers = new List<IEventApplier<TAggregate, TIdentity>>();
+
+        protected void Register(IEventApplier<TAggregate, TIdentity> eventApplier)
+        {
+            _eventAppliers.Add(eventApplier);
+        }
+
         private static readonly ConcurrentDictionary<Type, ConcurrentDictionary<Type, Action<TAggregate, IAggregateEvent>>> ApplyMethods = new ConcurrentDictionary<Type, ConcurrentDictionary<Type, Action<TAggregate, IAggregateEvent>>>(); 
-        private Action<TAggregate, IAggregateEvent> GetApplyMethod(Type aggregateEventType)
+        protected Action<TAggregate, IAggregateEvent> GetApplyMethod(Type aggregateEventType)
         {
             var aggregateType = GetType();
             var typeDictionary = ApplyMethods.GetOrAdd(
@@ -173,7 +178,12 @@ namespace EventFlow.Aggregates
                 aggregateEventType,
                 t =>
                     {
-                        var m = aggregateType.GetMethod("Apply", new []{ t });
+                        var m = aggregateType.GetMethod(
+                            "Apply",
+                            BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance,
+                            null,
+                            new []{ t },
+                            null);
                         if (m == null)
                         {
                             throw WrongImplementationException.With(
@@ -190,11 +200,7 @@ namespace EventFlow.Aggregates
 
         public override string ToString()
         {
-            return string.Format(
-                "{0} v{1}(-{2})",
-                GetType().Name,
-                Version,
-                _uncommittedEvents.Count);
+            return $"{GetType().Name} v{Version}(-{_uncommittedEvents.Count})";
         }
     }
 }