+++ RELEASE_NOTES.md
@@ -2,6 +2,13 @@
 
  * Breaking: `EventFlowOptions AddDefaults(...)` now also adds event
    definitions
+ * New: [RabbitMQ](http://www.rabbitmq.com/) is now supported through the new
+   NuGet package called `EventFlow.RabbitMQ` which enables domain events to be
+   published to the bus
+ * New: If you want to subscribe to all domain events, you can implement
+   and register a service that implements `ISubscribeSynchronousToAll`. Services
+   that implement this will automatically be added using the
+   `AddSubscribers(...)` or `AddDefaults(...)` extension to `EventFlowOptions`
 
 ### New in 0.10.642 (released 2015-08-17)